@@ -1,5 +1,6 @@
 package com.control.yape.service;
 
+import com.control.yape.dto.YapeNotificacionRequest;
 import com.control.yape.model.YapeMovimiento;
 
 import java.math.BigDecimal;
@@ -8,7 +9,7 @@ import java.util.List;
 
 public interface YapeMovimientoService {
 
-    List<YapeMovimiento> listarPorCuenta(Long cuentaId);
+	List<YapeMovimiento> listarPorCuenta(Long cuentaId);
 
     List<YapeMovimiento> listarPorEmpresaYRangoFechas(
             Long empresaId,
@@ -19,4 +20,15 @@ public interface YapeMovimientoService {
     YapeMovimiento registrarMovimiento(Long cuentaId, YapeMovimiento movimiento);
 
     BigDecimal totalRecibidoEmpresa(Long empresaId, LocalDateTime desde, LocalDateTime hasta);
+    
+    List<YapeMovimiento> listarTodos();
+    
+    List<YapeMovimiento> listarPorEmpresa(Long empresaId);
+    
+    YapeMovimiento registrarDesdeApp(YapeNotificacionRequest req, boolean[] duplicadoFlag);
+
+    // ✅ VALIDAR QUE LA CUENTA PERTENECE A UNA EMPRESA
+    void validarCuentaPertenece(Long cuentaId, Long empresaId);
 }
+
+
